Sequence point fix for the MSG size check macros (#5)
The first implementation of those macros were not safe against things
like post-increment on the index variable, which is used in the ADI
remap function in "application_data_object.c".

Local help variables has been added so that the macro arguments that
must be used multiple times only are 'executed' once.